+++ src/com/intellij/idea/plugin/hybris/system/type/file/TSDeploymentModelDomFileDescription.kt
@@ -1,0 +1,45 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.intellij.idea.plugin.hybris.system.type.file
+
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
+import com.intellij.idea.plugin.hybris.settings.components.ProjectSettingsComponent
+import com.intellij.idea.plugin.hybris.system.type.model.deployment.Model
+import com.intellij.openapi.module.Module
+import com.intellij.openapi.util.Iconable.IconFlags
+import com.intellij.psi.xml.XmlFile
+import com.intellij.util.xml.DomFileDescription
+import javax.swing.Icon
+
+class TSDeploymentModelDomFileDescription : DomFileDescription<Model>(Model::class.java, HybrisConstants.ROOT_TAG_DEPLOYMENT_MODEL_XML) {
+
+    override fun getFileIcon(@IconFlags flags: Int): Icon = HybrisIcons.TypeSystem.DEPLOYMENT_MODEL
+
+    override fun isMyFile(
+        file: XmlFile, module: Module?
+    ) = super.isMyFile(file, module)
+        && HybrisConstants.CORE_ADVANCED_DEPLOYMENT_FILE == file.name
+        && hasName(file)
+        && ProjectSettingsComponent.getInstance(file.project).isHybrisProject()
+
+    private fun hasName(file: XmlFile) = file.rootTag
+        ?.attributes
+        ?.any { it.localName == "name" && it.value == "hybris" }
+        ?: false
+}